REBASE: trying to fix semantics
