--- assets/enviosimples.js
@@ -1,3 +1,0 @@
-jQuery(document).ready(function(){
-
-});
