+++ es-plugin-woocommerce/.idea/encodings.xml
