+++ es-plugin-woocommerce/admin/css/es-plugin-woocommerce-admin.css
@@ -1,0 +1,4 @@
+/**
+ * All of the CSS for your admin-specific functionality should be
+ * included in this file.
+ */
