+++ es-plugin-woocommerce/admin/partials/es-plugin-woocommerce-admin-display.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Provide a admin area view for the plugin
+ *
+ * This file is used to markup the admin-facing aspects of the plugin.
+ *
+ * @link       https://github.com/Envio-Simples/es-plugin-woocommerce
+ * @since      1.0.0
+ *
+ * @package    Es_Plugin_Woocommerce
+ * @subpackage Es_Plugin_Woocommerce/admin/partials
+ */
+?>
+
+<!-- This file should primarily consist of HTML with a little bit of PHP. -->
