+++ es-plugin-woocommerce/includes/class-es-plugin-woocommerce-i18n.php
@@ -1,0 +1,47 @@
+<?php
+
+/**
+ * Define the internationalization functionality
+ *
+ * Loads and defines the internationalization files for this plugin
+ * so that it is ready for translation.
+ *
+ * @link       https://github.com/Envio-Simples/es-plugin-woocommerce
+ * @since      1.0.0
+ *
+ * @package    Es_Plugin_Woocommerce
+ * @subpackage Es_Plugin_Woocommerce/includes
+ */
+
+/**
+ * Define the internationalization functionality.
+ *
+ * Loads and defines the internationalization files for this plugin
+ * so that it is ready for translation.
+ *
+ * @since      1.0.0
+ * @package    Es_Plugin_Woocommerce
+ * @subpackage Es_Plugin_Woocommerce/includes
+ * @author     https://github.com/Envio-Simples/es-plugin-woocommerce <contato@ecomd.com.br>
+ */
+class Es_Plugin_Woocommerce_i18n {
+
+
+	/**
+	 * Load the plugin text domain for translation.
+	 *
+	 * @since    1.0.0
+	 */
+	public function load_plugin_textdomain() {
+
+		load_plugin_textdomain(
+			'es-plugin-woocommerce',
+			false,
+			dirname( dirname( plugin_basename( __FILE__ ) ) ) . '/languages/'
+		);
+
+	}
+
+
+
+}
