+++ es-plugin-woocommerce/index.php
@@ -1,0 +1,1 @@
+<?php // Silence is golden
