+++ es-plugin-woocommerce/public/css/es-plugin-woocommerce-public.css
@@ -1,0 +1,4 @@
+/**
+ * All of the CSS for your public-facing functionality should be
+ * included in this file.
+ */
