+++ es-plugin-woocommerce/public/index.php
@@ -1,0 +1,1 @@
+<?php // Silence is golden
