+++ es-plugin-woocommerce/public/js/enviobrudam.js
@@ -1,0 +1,3 @@
+jQuery(document).ready(function(){
+
+});
