+++ es-plugin-woocommerce/public/js/enviosimples.js
@@ -1,0 +1,3 @@
+jQuery(document).ready(function(){
+
+});
